+++ packages/@rrios-dev__react-formik-checkbox/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rrios-dev/react-formik-checkbox
 
+## 0.0.3
+
+### Patch Changes
+
+- Add fieldset component and add .gitignore
+- Updated dependencies
+  - @rrios-dev/react-form-checkbox@0.0.5
+
 ## 0.0.2
 
 ### Patch Changes