+++ packages/@rrios-dev__react-formik-radio/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rrios-dev/react-formik-radio
 
+## 0.0.3
+
+### Patch Changes
+
+- Add fieldset component and add .gitignore
+- Updated dependencies
+  - @rrios-dev/react-form-radio@0.0.3
+
 ## 0.0.2
 
 ### Patch Changes