+++ packages/@rrios-dev__react-formik-select/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rrios-dev/react-formik-select
 
+## 0.0.3
+
+### Patch Changes
+
+- Add fieldset component and add .gitignore
+- Updated dependencies
+  - @rrios-dev/react-form-select@0.0.5
+
 ## 0.0.2
 
 ### Patch Changes