+++ packages/@rrios-dev__react-text/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rrios-dev/react-text
 
+## 0.0.4
+
+### Patch Changes
+
+- Add fieldset component and add .gitignore
+- Updated dependencies
+  - @rrios-dev/styled@0.0.5
+
 ## 0.0.3
 
 ### Patch Changes